SmallDatasetMaker

Documentation for SmallDatasetMaker.

SmallDatasetMaker.DATASET_ABS_DIR

SmallDatasetMaker.SourceData

SmallDatasetMaker.SourceData

SmallDatasetMaker.SourceData

SmallDatasetMaker.SourceData

SmallDatasetMaker.SourceData

SmallDatasetMaker.SourceData

SmallDatasetMaker.SourceData

SmallDatasetMaker.abspath

SmallDatasetMaker.compress_save

SmallDatasetMaker.compress_save!

SmallDatasetMaker.dataset

SmallDatasetMaker.dataset

SmallDatasetMaker.dataset_dir

SmallDatasetMaker.dataset_table

SmallDatasetMaker.datasets

SmallDatasetMaker.dir_data

SmallDatasetMaker.dir_raw

SmallDatasetMaker.dir_to_be_converted

SmallDatasetMaker.get_package_dataset_name

SmallDatasetMaker.load_original

SmallDatasetMaker.return_compressed

SmallDatasetMaker.return_compressed

SmallDatasetMaker.target_row

SmallDatasetMaker.unzip_file

SmallDatasetMaker.unzip_file

SmallDatasetMaker.DATASET_ABS_DIR — Constant

Absolute directory for package SmallDatasetMaker. Also see abspath.

SmallDatasetMaker.SourceData — Method

If zipfile not specified, it will be dir_data(package_name, dataset_name*".gz").

SmallDatasetMaker.SourceData — Method

If rows, columns not specified, CSV.read(srcfile, DataFrame) will be applied to get the number of rows/columns.

SmallDatasetMaker.SourceData — Method

If description not specified, it will be "".

SmallDatasetMaker.SourceData — Method

SourceData(srcfile, packagename, datasetname, title, zipfile, rows, columns, description, timestamps)

srcfile is the path to the source file, the package_name will be the folder that the file resides, the dataset_name will be the name of the data without extension.

If timestamps not specified, it will be today().

SmallDatasetMaker.SourceData — Method

If title not specified, it will be "Data [$dataset_name] of [$package_name]".

SmallDatasetMaker.SourceData — Method

If package_name, dataset_name not specified, (package_name, dataset_name) = get_package_dataset_name(srcfile) is applied.

SmallDatasetMaker.SourceData — Method

SourceData(row::DataFrameRow) returns

SourceData(
    abspath(row.RawData),    # srcfile::Union{Missing,String}
    row.PackageName,         # package_name::String
    row.Dataset,             # dataset_name::String
    row.Title,               # title::Union{Missing,String}
    abspath(row.ZippedData), # zipfile::String
    row.Rows,                # rows::Int
    row.Columns,             # columns::Int
    row.Description,         # description::Union{Missing,String}
    row.TimeStamp,           # timestamps::TimeType
    )

SmallDatasetMaker.abspath — Method

abspath(args...) = joinpath(DATASET_ABS_DIR[], args...)

SmallDatasetMaker.compress_save! — Method

compress_save!(SD::SourceData; move_source = true) compress the SD.srcfile, save the zipped one to SD.zipfile, and update the /home/runner/work/SmallDatasetMaker.jl/SmallDatasetMaker.jl/data/doc/datasets.csv. By default, move_source = true that the source file will be moved to dir_raw().

SmallDatasetMaker.compress_save — Method

compress_save(srcpath) is equivalent to compress_save!(SourceData(srcpath)) but returns SD::SourceData.

compress_save takes the same keyword arguments as compress_save!.

SmallDatasetMaker.dataset — Method

dataset(package_name::AbstractString, dataset_name::AbstractString) returns a DataFrame object unzipped from the last row returned by target_row(package_name, dataset_name). This function mimics the dataset function in RDatasets.jl.

SmallDatasetMaker.dataset — Method

dataset(target_path) decompress target_path and returns it as a DataFrame.

Notice!

If you were intended to load target_path under SmallDatasetMaker rather than the current directory your working with, you should apply abspath that target_path = SmallDatasetMaker.abspath(target_path).

SmallDatasetMaker.dataset_dir — Method

Absolute path to the "data" of SmallDatasetMaker. See also DATASET_ABS_DIR.

Example

`dataset_dir("doc", "datasets.csv")`

# Output
"~/.julia/dev/SmallDatasetMaker/data/doc/datasets.csv"

SmallDatasetMaker.dataset_table — Method

The path to the index table for datasets in SmallDatasetMaker. If SmallDatasetMaker is added using pkg> dev SmallDatasetMaker in other project/environment, dataset_table() returns "~/.julia/dev/SmallDatasetMaker/src/../data/doc/datasets.csv".

The reason for dataset_table to be a function rather than a constant is that I can redefine it in the scope of test. See test/compdecomp.jl.

SmallDatasetMaker.datasets — Method

datasets() returns the table of this dataset, and define the global variable SmallDatasetMaker.__datasets as this table.

Set ; update_table = true to force update SmallDatasetMaker.__datasets with the dataset_table(); this keyword argument is intended to make some tests can work since in test dataset_table() is mutating. # todo: find a better way to deal with it.

SmallDatasetMaker.dir_data — Method

Path to the directory for the data.

SmallDatasetMaker.dir_raw — Method

Path to the directory for the backup of the raw data; only for package internal use.

SmallDatasetMaker.dir_to_be_converted — Method

Path to the directory for the to-be-compressed data; only for package internal use.

SmallDatasetMaker.get_package_dataset_name — Method

Given path to the source file, get_package_dataset_name(srcpath) derive package name and dataset name from the srcpath.

Example

srcpath = joinpath("Whatever", "RDatasets", "iris.csv")
SmallDatasetMaker.get_package_dataset_name(srcpath)

# output

("RDatasets", "iris")

SmallDatasetMaker.load_original — Method

load_original(path::AbstractString) opens path and return the read data.

SmallDatasetMaker.return_compressed — Method

return_compressed(path::AbstractString) returned compressed data.

Example

compressed = return_compressed("data/data.csv")

SmallDatasetMaker.return_compressed — Method

return_compressed(data::Vector{UInt8}) returned compressed data.

Example

data = load_original("data/data.csv")
compressed = return_compressed(data)

SmallDatasetMaker.target_row — Method

target_row returns the latest information in datasets(). Given package_name, dataset_name, target_row(package_name, dataset_name), target_row returns the last row that matches row.PackageName == package_name && row.Dataset == dataset_name".

SmallDatasetMaker.unzip_file — Method

The same as dataset, but also save the unzip file.

SmallDatasetMaker.unzip_file — Method

unzip_file(target_path) unzip file at target_path to current directory preserve its original name.

Notice!

If you were intended to load target_path under SmallDatasetMaker rather than the current directory your working with, you should apply abspath that target_path = SmallDatasetMaker.abspath(target_path).